\name{createCOMObject-methods}
\docType{methods}
\alias{createCOMObject,COMEventServerInfo-method}
\title{Create DCOM server instance from information about an event interface}
\description{
This function creates a DCOM server instance using
information generated from a type library about an
event interface. The information provides descriptions
of the methods expected by the server along with information
about the method identifiers, etc.
}
\section{Methods}{
\describe{
\item{def = "COMEventServerInfo"}{creates an server instance from the event handler description}
}}
\keyword{methods}
